USC-SIkJalpha submitting Round 3
I removed SK, CH, and GB due to anomalous projections. Particularly, for CH and GB, I don't have the booster data and I did not have time to integrate from other sources.
For the countries where 2nd booster data is available, I copied the same adoption trends for all future boosters.
For countries where it was not available, I copied the 1st booster adoption trend. As a result for these countries, the impact of the future may be higher and you may see a flatter projection.
Again, I think it was possible to somehow model a reduction in uptake in the future from the first booster. But I did not have time to do it.
Anyway, I hope these projections are helpful.