Update submission for Round 3 UC3M-EpiGraph
We have updated our previous submission with three more samples and a different approach for counting available booster doses.
For the extra booster campaign, we assume that 50% of the population will receive the booster. This means that 43% of the population 70+ will receive a booster and only 23.8% of the population between 20 and 60 years will be vaccinated with boosters. Note that these small coverages will limit the effectiveness of the vaccination campaign.