Correction negative values
We have corrected the negative values in sample 16, scenario A, child and inc death. Indeed, values for "child" should be null for every variable except "inc infection". However, due to the interpolation done when calculating the values, negative numbers can appear. We have corrected this and set those values to 0.